+++ src/main/kotlin/org/jetbrains/intellij/dependency/PluginDependencyImpl.kt
@@ -17,9 +17,9 @@ class PluginDependencyImpl(
 
     override var channel: String? = null
     override var jarFiles: Collection<File> = emptyList()
+    override var sourceJarFiles: Collection<File> = emptyList()
     override var classesDirectory: File? = null
     override var metaInfDirectory: File? = null
-    override val sourcesDirectory: File? = null
     override val notation = PluginDependencyNotation(id, version, channel)
     var sinceBuild: String? = null
     var untilBuild: String? = null
@@ -32,6 +32,7 @@ class PluginDependencyImpl(
             val lib = File(artifact, "lib")
             if (lib.isDirectory) {
                 jarFiles = collectJars(lib)
+                sourceJarFiles = collectJars(File(lib, "src"))
             }
             val classes = File(artifact, "classes")
             if (classes.isDirectory) {
@@ -62,9 +63,9 @@ class PluginDependencyImpl(
         if (maven != other.maven) return false
         if (channel != other.channel) return false
         if (jarFiles != other.jarFiles) return false
+        if (sourceJarFiles != other.sourceJarFiles) return false
         if (classesDirectory != other.classesDirectory) return false
         if (metaInfDirectory != other.metaInfDirectory) return false
-        if (sourcesDirectory != other.sourcesDirectory) return false
         if (notation != other.notation) return false
         if (sinceBuild != other.sinceBuild) return false
         if (untilBuild != other.untilBuild) return false
@@ -80,9 +81,9 @@ class PluginDependencyImpl(
         result = 31 * result + maven.hashCode()
         result = 31 * result + (channel?.hashCode() ?: 0)
         result = 31 * result + jarFiles.hashCode()
+        result = 31 * result + sourceJarFiles.hashCode()
         result = 31 * result + (classesDirectory?.hashCode() ?: 0)
         result = 31 * result + (metaInfDirectory?.hashCode() ?: 0)
-        result = 31 * result + (sourcesDirectory?.hashCode() ?: 0)
         result = 31 * result + notation.hashCode()
         result = 31 * result + (sinceBuild?.hashCode() ?: 0)
         result = 31 * result + (untilBuild?.hashCode() ?: 0)