+++ src/main/kotlin/org/jetbrains/intellij/dependency/PluginProjectDependency.kt
@@ -44,14 +44,14 @@ class PluginProjectDependency(private val pluginDirectory: File, val context: St
     override val jarFiles: Collection<File>
         get() = pluginDependency?.jarFiles ?: emptyList()
 
+    override val sourceJarFiles: Collection<File>
+        get() = pluginDependency?.sourceJarFiles ?: emptyList()
+
     override val classesDirectory: File?
         get() = pluginDependency?.classesDirectory
 
     override val metaInfDirectory: File?
         get() = pluginDependency?.metaInfDirectory
-
-    override val sourcesDirectory: File?
-        get() = pluginDependency?.sourcesDirectory
 
     override val builtin: Boolean
         get() = false