+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpecBase.kt
@@ -21,10 +21,10 @@ import kotlin.test.assertEquals
 abstract class IntelliJPluginSpecBase {
 
     private var debugEnabled = true
-    private val kotlinPluginVersion = System.getProperty("test.kotlin.version")
     private val gradleDefault = System.getProperty("test.gradle.default")
     private val gradleArguments = System.getProperty("test.gradle.arguments", "")
         .split(' ').filter(String::isNotEmpty).toTypedArray()
+    protected val kotlinPluginVersion: String = System.getProperty("test.kotlin.version")
     protected val gradleVersion: String = System.getProperty("test.gradle.version").takeUnless { it.isNullOrEmpty() } ?: gradleDefault
 
     val gradleHome: String = System.getProperty("test.gradle.home")
@@ -76,7 +76,7 @@ abstract class IntelliJPluginSpecBase {
         """
         )
 
-        gradleProperties.groovy(
+        gradleProperties.properties(
             """
             kotlin.stdlib.default.dependency = false
         """
@@ -236,11 +236,16 @@ abstract class IntelliJPluginSpecBase {
     // https://youtrack.jetbrains.com/issue/KTIJ-1001
     fun File.xml(@Language("XML") content: String) = append(content)
 
+    /**
+     * Appends the passed content to this file. Overwrites existing content if the `overwrite` parameter is true.
+     */
     fun File.groovy(@Language("Groovy") content: String) = append(content)
 
     fun File.java(@Language("Java") content: String) = append(content)
 
     fun File.kotlin(@Language("kotlin") content: String) = append(content)
+
+    fun File.properties(@Language("Properties") content: String) = append(content)
 
     private fun File.append(content: String) = appendText(content.trimIndent() + "\n")
 }