+++ CHANGES.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## next
+### Added
+- Automatically detect bundled sources in plugin dependency
+
 ### Fixed
 - Fixed broken instrumentation when custom sources directory is set [#1004](../../issues/1004)
 - Fixed `java.nio.file.FileAlreadyExistsException: instrumentCode/classpath.index` [#998](../../issues/998)